feat!: rewrite modules/files using TypeScript
