--- packages/micro/lib/error.js
@@ -1,4 +1,0 @@
-module.exports = (message, errorCode) => {
-	console.error(`micro: ${message}`);
-	console.error(`micro: https://err.sh/micro/${errorCode}`);
-};
