+++ packages/micro/package.json
@@ -34,9 +34,12 @@
   "dependencies": {
     "arg": "4.1.0",
     "content-type": "1.0.4",
-    "raw-body": "2.4.1"
+    "raw-body": "2.4.1",
+    "tsimportlib": "0.0.3"
   },
   "devDependencies": {
+    "@types/content-type": "1.1.5",
+    "@types/node": "18.0.3",
     "@vercel/style-guide": "3.0.0",
     "eslint": "8.19.0",
     "prettier": "2.7.1",
